Move to_regex to mathics.core.eval...
and deduce hoakiness in filesystem.py
